irfanhassan_skipq2021/Sprint3_irfan/requirements.txt
aws-cdk.core==1.135.0
aws-cdk.assertions==1.135.0
aws-cdk.assertions==1.135.0
aws-cdk.aws.lambda==1.135.0
aws-cdk.aws_events_targets==1.135.0
aws-cdk.aws_dynamodb==1.135.0
aws-cdk.aws_sns_subscriptions==1.135.0
aws-cdk.aws_cloudwatch_actions==1.135.0
aws-cdk.aws_s3==1.135.0
aws-cdk.aws_lambda_event_sources==1.135.0
aws-cdk.aws_sns==1.135.0
aws-cdk.aws_iam==1.135.0
aws-cdk.aws_events==1.135.0
aws-cdk.aws_s3_notifications==1.135.0
aws-cdk.pipelines==1.135.0
aws-cdk.aws_codepipeline_actions==1.135.0
aws-cdk.aws_codedeploy==1.135.0
aws-cdk.aws_apigateway==1.135.0
pytest
boto3
requests